+++ tracer/test/Datadog.Trace.Security.IntegrationTests/AspNetCore2.cs
@@ -9,7 +9,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
-using Datadog.Trace.IntegrationTestHelpers;
+using Datadog.Trace.TestHelpers;
 using Xunit;
 using Xunit.Abstractions;
