+++ tracer/test/Datadog.Trace.Security.IntegrationTests/AspNetCore5ExternalRules.cs
@@ -5,7 +5,7 @@
 
 // The conditions looks weird, but it seems like _OR_GREATER is not supported yet in all environments
 // We can trim all the additional conditions when this is fixed
-#if NETCOREAPP3_0 || NETCOREAPP3_1 || NET5_0 || NET6_0
+#if NETCOREAPP3_0_OR_GREATER
 
 using System;
 using System.Collections.Generic;
@@ -13,7 +13,6 @@ using System.IO;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
-using Datadog.Trace.IntegrationTestHelpers;
 using Datadog.Trace.TestHelpers;
 using Xunit;
 using Xunit.Abstractions;
